Fluor obtainable bis aquatisiertes Fluor
